+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Enh #354: Refactor PHP type of `ColumnSchemaInterface` instances (@Tigrov)
 - Enh #356: Raise minimum PHP version to `^8.1` with minor refactoring (@Tigrov)
 - Enh #355: Implement `ColumnFactory` class (@Tigrov)
+- Enh #359: Separate column type constants (@Tigrov)
+- Enh #359: Remove `Schema::TYPE_ARRAY` and `Schema::TYPE_STRUCTURED` constants (@Tigrov)
 
 ## 1.3.0 March 21, 2024